[YUNIKORN-1375] [shim] Pass group info from the shim to the core (#470)
Closes: #470
