+++ pkg/plugin/admissioncontrollers/webhook/admission_controller.go
@@ -23,6 +23,7 @@ import (
 	"crypto/sha256"
 	"encoding/json"
 	"fmt"
+	"github.com/apache/yunikorn-k8shim/pkg/common/utils"
 	"io"
 	"net/http"
 	"os"
@@ -240,7 +241,7 @@ func (c *admissionController) mutate(req *admissionv1.AdmissionRequest) *admissi
 		return admissionResponseBuilder(uid, false, err.Error(), nil)
 	}
 
-	if annotation, ok := pod.Annotations[userInfoAnnotation]; ok && !c.bypassAuth {
+	if annotation := utils.GetPodAnnotationValue(&pod, userInfoAnnotation); annotation != "" && !c.bypassAuth {
 		userName := req.UserInfo.Username
 		groups := req.UserInfo.Groups
 
@@ -259,7 +260,7 @@ func (c *admissionController) mutate(req *admissionv1.AdmissionRequest) *admissi
 		}
 	}
 
-	if labelAppValue, ok := pod.Labels[constants.LabelApp]; ok {
+	if labelAppValue := utils.GetPodLabelValue(&pod, constants.LabelApp); labelAppValue != "" {
 		if labelAppValue == yunikornPod {
 			log.Logger().Info("ignore yunikorn pod")
 			return admissionResponseBuilder(uid, true, "", nil)
@@ -318,30 +319,7 @@ func updateLabels(namespace string, pod *v1.Pod, patch []patchOperation) []patch
 		zap.String("namespace", namespace),
 		zap.Any("labels", pod.Labels))
 
-	existingLabels := pod.Labels
-	result := make(map[string]string)
-	for k, v := range existingLabels {
-		result[k] = v
-	}
-
-	if _, ok := existingLabels[constants.SparkLabelAppID]; !ok {
-		if _, ok := existingLabels[constants.LabelApplicationID]; !ok {
-			// if app id not exist, generate one
-			// for each namespace, we group unnamed pods to one single app
-			// application ID convention: ${AUTO_GEN_PREFIX}-${NAMESPACE}-${AUTO_GEN_SUFFIX}
-			generatedID := generateAppID(namespace)
-			result[constants.LabelApplicationID] = generatedID
-
-			// if we generate an app ID, disable state-aware scheduling for this app
-			if _, ok := existingLabels[constants.LabelDisableStateAware]; !ok {
-				result[constants.LabelDisableStateAware] = "true"
-			}
-		}
-	}
-
-	if _, ok := existingLabels[constants.LabelQueueName]; !ok {
-		result[constants.LabelQueueName] = defaultQueue
-	}
+	result := utils.UpdatePodLabelForAdmissionController(pod, namespace)
 
 	patch = append(patch, patchOperation{
 		Op:    "add",
